fixed VFM, changed includes

+++ libc/string/strlen.c
@@ -1,6 +1,6 @@
 #include "../include/string.h"
 
-int strlen(const char* str){
+unsigned long strlen(const char* str){
     size_t len = 0;
     while (str[len])
         ++len;
